+++ src/lib/Region.cpp
@@ -6,7 +6,7 @@
 //-----------------------------------------------------------------------------
 // Region
 //-----------------------------------------------------------------------------
-Region::Region(UInt16 addrTop) :
+Region::Region(Integer addrTop) :
 	_cntRef(1), _addrTop(addrTop), _pRegionsIngredient(new RegionOwner())
 {
 	_buff.reserve(4096);
